refact: schedule service 리팩터링 - 함수 분리, 폴더명 변경
